requirements.txt
fastapi
uvicorn
sqlmodel
pymysql
python-dotenv
passlib[bcrypt]
python-jose
redis
pydantic-settings
